number_guessing_wb/requirements.txt
flask
flask-Cors
Pillow
numpy
torch
torchvision